refactor(app): extract overlay state into app/overlays.rs
ActionItem, PendingConfirm, ActionMenu, BranchCreateInput, and
Overlays move verbatim, together with the text-edit helpers
(insert_char_at / remove_char_at, now pub(super) for the input key
handler in mod.rs; char_to_byte_idx stays file-private) and their
tests. Re-exported from app. Part of #270.
